Replace use of getty with some Python script.
getty is now more picky than it was and does not accept to turn on job
control on our console. We replace it with some Python wrapper that
does the job. Unfortunately, it seems that tcsetpgrp(0,1) does not
work and therefore we replace zsh by bash which calls this function
(and it works!).

Getting job control seems a complicated matter...