arista-veos: use two different VLAN to Internet to avoid any switching
This should not be necessary, but I am too lazy to test. The rationale
behind this is that EOS may share information with its peer if it
believes that this is the same VLAN. We use VLAN only because it was not
possible to use raw Ethernet interface. However, VLAN are declared as
switchport. We don't really want that.