l3-hyperv: disable redirects
As a router, we already don't accept them. However, we should not send
them either. It takes time for redirects to expire and we prefer to
route on the behalf of the VM instead of doing redirect. Unfortunately,
this setting is not available for IPv6.

RFC4294 (4.2) says: "If the node is a router, Redirect functionality
MUST be supported". Unfortunate.