l3-hyperv: spawn DNS server on each hyperv
Instead of doing a relay, just spawn a DNS server on each hyperv. This
will help future evolutions and match what we intend to do. The
anycasted DHCP would have been interesting to do too, but it would need
to live on public IP space and we don't have public IP to spare.