l3-hyperv: tentative to work without logical systems
Those are not supported on QFX. Instead, try to use
rib-groups. Unfortunately, this doesn't work either:

    BGP: ribgroup public-v4: public-v4.inet.0 not a valid secondary rib for this nlri.

It seems that a secondary RIB has to be a primary RIB in another VRF or
something like that. Dunno why I cannot just create and specify a RIB.

Also, if it worked, we have several other steps to ensure:

 - Are all routes eligible to be copied to secondary RIB or only the
   best ones? We want all routes to be able to have add-path (this is
   the primary purpose of this setup).

 - We need to ensure that we don't turn the device into a router. Routes
   are not installed, but we still mix inside the same routing table the
   private and public device routes. A filter should be installed on
   each interface to ensure packets don't come from a forbidden
   interface.

 - If all routes are also in inet.0, we need to ensure that we don't
   waste memory on this. Maybe structs are shared. Maybe not.