Revert "l3-hyperv: use of suppress_ifgroup"
This reverts commit c8fbcbdbf053109bd6e52a5134e541aecb63e744.
"suppress_ifgroup" doesn't work this way. A lookup is done and if the
lookup provides a destination whose group is the one specified, the
result is suppressed. It would be a way to use the same routing table
for both public and private interfaces.

See Linux commit 7764a45a8f1fe74d4f7d301eaca2e558e7e2831a.