common: update SMBIOS values for vMX
QFX and VRR are likely to have to be updated as well, but I don't know
the values. They were found in the libvirt templates shipped with vMX.
The content of loader.conf seems to be still correct but there seems
to be a lot more now. Maybe we should use the script provided by
Juniper to see what it does.